gcs: refactor rasterizer and overall clean up
The rough tile rasterizer loop was replaced in
favor of sorted bounding box edge lists. This
greatly improves perf in cases where primitives
don't cover large portions of the rasterized area
by skipping tiles which are not covered by their
bboxes.

This comes at a cost of setting up the edge lists
in raster setup but the penalty is neglegable
compared to the time saved not iterating empty tiles.